requirements.txt
imgaug==0.4.0
pandas==1.4.0
Pillow==9.0.1
scikit-image==0.18.3
scipy==1.7.1
torchvision==0.10.1
torch==1.9.1
git+https://github.com/npielawski/octopytorch.git@a59e74b809c1a2734903ad6447bfbf591794de2e